fix(*): Start server only on machines without display
