+++ src/General/MB.java
@@ -2,7 +2,6 @@ package General;
 
 
 import General.Shared.MBPanel;
-import Server.Server;
 
 import javax.swing.*;
 import java.awt.*;
@@ -26,26 +25,16 @@ public class MB {
 
     /**
      * Setup the JFrame and show the menu
-     *
-     * @param args that were passed
      */
-    public static void main(String[] args) {
-        if (args.length > 0) {
-            if ("-s".equals(args[0])) {
-                new Server("Multi-Bomb server").run();
-            } else {
-                System.out.println("Option not recognized, try -s to start server only");
-            }
-        } else {
-            MB.settings.loadSettings();
-            setupFrame();
-            show(new Intro(), false);
+    public static void startGame() {
+        MB.settings.loadSettings();
+        setupFrame();
+        show(new Intro(), false);
 
-            // Set fullscreen if necessary
-            if (settings.fullscreen) {
-                Window activeWindow = javax.swing.FocusManager.getCurrentManager().getActiveWindow();
-                activeWindow.getGraphicsConfiguration().getDevice().setFullScreenWindow(MB.frame);
-            }
+        // Set fullscreen if necessary
+        if (settings.fullscreen) {
+            Window activeWindow = javax.swing.FocusManager.getCurrentManager().getActiveWindow();
+            activeWindow.getGraphicsConfiguration().getDevice().setFullScreenWindow(MB.frame);
         }
     }
